+++ site/dat/docs/Changelog.md
@@ -3,6 +3,7 @@
  - add worker id to cloud log file names
  - add `cloud` and `local` aliases 
  - fix delay trouble in provisioning
+ - fix Gatling keepalive type cast bug
 
 ## 1.6.6 <sup>08 aug 2016</sup>
  - optimize aggregator by removing excessive calls to `BetterDict.get()`